pyproject.toml

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "refit-ai-service"
version = "0.1.0"
description = "Re-Fit AI Service API"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "beautifulsoup4>=4.12.0",
    "boto3>=1.28.0",
    "fastapi>=0.109.0",
    "google-genai>=1.0.0",
    "httpx>=0.27.0",
    "lxml>=5.0.0",
    "pgvector>=0.2.0",
    "presidio-analyzer>=2.2.0",
    "presidio-anonymizer>=2.2.0",
    "psycopg2-binary>=2.9.0",
    "pydantic>=2.0.0",
    "pymupdf>=1.24.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.6",
    "sentence-transformers>=2.2.0",
    "sqlalchemy>=2.0.0",
    "torch>=2.0.0",
    "transformers>=4.30.0",
    "uvicorn[standard]>=0.27.0",
    # 의존성 충돌 방지: grpcio 버전 고정
    "grpcio>=1.60.0,<1.70.0",
    "grpcio-status>=1.60.0,<1.70.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "bandit>=1.7.0",
    "build>=1.0.0",
]

[tool.setuptools]
include-package-data = true  # MANIFEST.in 사용

[tool.setuptools.packages.find]
where = ["."]
include = ["ai_app*"]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false